+++ src/connectors.rs
@@ -1264,6 +1264,7 @@ pub(crate) fn builtin_connector_types() -> Vec<Box<dyn ConnectorBuilder + 'stati
         Box::new(impls::gpubsub::consumer::Builder::default()),
         Box::new(impls::gpubsub::producer::Builder::default()),
         Box::new(impls::clickhouse::Builder::default()),
+        Box::new(impls::gcl::writer::Builder::default()),
     ]
 }